# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: revenue
    description: Revenue operations
  - name: sync
    description: Data sync operations
  - name: metrics
    description: Metric registry and typed query endpoint
  - name: financial-statements
    description: Financial statement operations
  - name: companies
    description: Company / governance read endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /revenue:
    get:
      operationId: getRevenue
      tags: [revenue]
      summary: List revenue records
      description: Returns revenue records with optional filtering by company, year, and month
      parameters:
        - name: company_id
          in: query
          required: false
          schema:
            type: string
        - name: year
          in: query
          required: false
          schema:
            type: integer
        - name: month
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Revenue records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueRecord"
  /financial-statements:
    get:
      operationId: getFinancialStatements
      tags: [financial-statements]
      summary: Fetch financial statements for a company and fiscal year
      description: Returns the income statement and balance sheet (or one if `type` is given) for the specified company and fiscal year, as a hierarchical list of lines.
      parameters:
        - name: company_id
          in: query
          required: true
          schema:
            type: string
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [income_statement, balance_sheet]
      responses:
        "200":
          description: Financial statements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialStatementsResponse"
    post:
      operationId: saveFinancialStatement
      tags: [financial-statements]
      summary: Upsert a parsed financial statement (replaces existing rows for same company/year/type)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialStatementSaveRequest"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialStatementSaveResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryError"
        "422":
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialStatementSaveError"
  /financial-statements/parse-pdf:
    post:
      operationId: parseFinancialStatementPdf
      tags: [financial-statements]
      summary: Convert an uploaded PDF financial statement into markdown
      description: |
        Streams the request body (raw PDF, content-type `application/pdf`) to LlamaParse
        (LlamaCloud) and returns the parsed markdown. The markdown can then be sent to
        `/financial-statements/preview` and `/financial-statements`.
      requestBody:
        required: true
        content:
          application/pdf:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Parsed markdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialStatementPdfMarkdown"
        "503":
          description: LlamaParse not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryError"
  /financial-statements/preview:
    post:
      operationId: previewFinancialStatement
      tags: [financial-statements]
      summary: Parse a markdown statement and return rows + verification issues without saving
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialStatementPreviewRequest"
      responses:
        "200":
          description: Preview rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialStatementPreview"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryError"
  /companies:
    get:
      operationId: getCompanies
      tags: [companies]
      summary: List companies with locations and shareholders
      description: Returns every company joined with its locations and ownership rows, in the shape consumed by the dashboard governance pages.
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyWithRelations"
  /family:
    get:
      operationId: getFamilyMembers
      tags: [companies]
      summary: List family members
      description: Returns the family member roster (with role and display color) used in the dashboard family bar.
      responses:
        "200":
          description: Family members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyMember"
  /sync/revenue:
    post:
      operationId: triggerRevenueSync
      tags: [sync]
      summary: Trigger revenue sync from Google Sheets
      description: Runs the Google Sheets revenue sync job immediately and returns the result
      responses:
        "200":
          description: Sync completed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "500":
          description: Sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
  /metrics:
    get:
      operationId: listMetrics
      tags: [metrics]
      summary: List metric definitions
      description: Returns the server-side metric registry so the frontend can render labels, units, and formats without duplicating them.
      responses:
        "200":
          description: Metric registry
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MetricDefinition"
  /query:
    post:
      operationId: runQuery
      tags: [metrics]
      summary: Run a typed metric query
      description: Resolves the requested metrics against the server-side registry and returns rows + column metadata. Replaces ad-hoc client-side aggregation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: Query result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "400":
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryError"
        "501":
          description: Dataset not yet available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SyncResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    FinancialStatementLine:
      type: object
      properties:
        id:
          type: string
        sort_order:
          type: integer
        depth:
          type: integer
        section_code:
          type: string
          nullable: true
        account_code:
          type: string
          nullable: true
        account_name_ko:
          type: string
        amount:
          type: string
          nullable: true
        prior_amount:
          type: string
          nullable: true
        is_subtotal:
          type: boolean
        parent_line_id:
          type: string
          nullable: true
        children:
          type: array
          items:
            $ref: "#/components/schemas/FinancialStatementLine"
      required:
        - id
        - sort_order
        - depth
        - account_name_ko
        - is_subtotal
        - children
    FinancialStatement:
      type: object
      properties:
        company_id:
          type: string
        fiscal_year:
          type: integer
        statement_type:
          type: string
        period_start:
          type: string
          nullable: true
        period_end:
          type: string
          nullable: true
        currency:
          type: string
        unit:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/FinancialStatementLine"
      required:
        - company_id
        - fiscal_year
        - statement_type
        - currency
        - unit
        - lines
    FinancialStatementsResponse:
      type: object
      properties:
        income_statement:
          oneOf:
            - $ref: "#/components/schemas/FinancialStatement"
            - type: "null"
        balance_sheet:
          oneOf:
            - $ref: "#/components/schemas/FinancialStatement"
            - type: "null"
      required:
        - income_statement
        - balance_sheet
    StatementVerifyIssue:
      type: object
      properties:
        period:
          type: string
          enum: [current, prior]
        message:
          type: string
      required:
        - period
        - message
    ParsedFinancialLine:
      type: object
      properties:
        account_name_ko:
          type: string
        section_code:
          type: string
          nullable: true
        depth:
          type: integer
        is_subtotal:
          type: boolean
        amount:
          type: string
          nullable: true
        prior_amount:
          type: string
          nullable: true
      required:
        - account_name_ko
        - depth
        - is_subtotal
    FinancialStatementPdfMarkdown:
      type: object
      properties:
        markdown:
          type: string
      required:
        - markdown
    FinancialStatementPreviewRequest:
      type: object
      properties:
        markdown:
          type: string
        statement_type:
          type: string
          enum: [income_statement, balance_sheet]
      required:
        - markdown
        - statement_type
    FinancialStatementPreview:
      type: object
      properties:
        statement_type:
          type: string
          enum: [income_statement, balance_sheet]
        line_count:
          type: integer
        lines:
          type: array
          items:
            $ref: "#/components/schemas/ParsedFinancialLine"
        issues:
          type: array
          items:
            $ref: "#/components/schemas/StatementVerifyIssue"
      required:
        - statement_type
        - line_count
        - lines
        - issues
    FinancialStatementSaveRequest:
      type: object
      properties:
        company_id:
          type: string
        fiscal_year:
          type: integer
        statement_type:
          type: string
          enum: [income_statement, balance_sheet]
        markdown:
          type: string
        period_start:
          type: string
          nullable: true
        period_end:
          type: string
          nullable: true
        skip_verification:
          type: boolean
          nullable: true
      required:
        - company_id
        - fiscal_year
        - statement_type
        - markdown
    FinancialStatementSaveResult:
      type: object
      properties:
        statement_id:
          type: string
        line_count:
          type: integer
        issues:
          type: array
          items:
            $ref: "#/components/schemas/StatementVerifyIssue"
      required:
        - statement_id
        - line_count
        - issues
    FinancialStatementSaveError:
      type: object
      properties:
        error:
          type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/StatementVerifyIssue"
      required:
        - error
        - issues
    Shareholder:
      type: object
      properties:
        id:
          type: string
        company_id:
          type: string
        name:
          type: string
        percentage:
          type: number
        is_entity:
          type: boolean
        updated_at:
          type: string
      required:
        - id
        - company_id
        - name
        - percentage
        - is_entity
        - updated_at
    CompanyWithRelations:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        short_name:
          type: string
          nullable: true
        category:
          type: string
        parent_id:
          type: string
          nullable: true
        locations:
          type: array
          items:
            type: string
        created_at:
          type: string
        shareholders:
          type: array
          items:
            $ref: "#/components/schemas/Shareholder"
        directors:
          type: array
          items:
            type: object
      required:
        - id
        - name
        - category
        - locations
        - created_at
        - shareholders
        - directors
    FamilyMember:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        color:
          type: string
      required:
        - name
        - role
        - color
    RevenueRecord:
      type: object
      properties:
        id:
          type: string
        company_id:
          type: string
        year:
          type: integer
        month:
          type: integer
        amount:
          type: string
        category:
          type: string
          nullable: true
        memo:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - company_id
        - year
        - month
        - amount
        - created_at
    MetricDefinition:
      type: object
      properties:
        code:
          type: string
        label_ko:
          type: string
        unit:
          type: string
          enum: [KRW, PCT, YEARS, COUNT]
        format:
          type: string
        dataset:
          type: string
          enum: [revenue, kpi]
        description:
          type: string
          nullable: true
      required:
        - code
        - label_ko
        - unit
        - format
        - dataset
    QueryFilter:
      type: object
      description: One filter clause. Use 'value' for eq/gte/lte, 'values' for in, and min/max for between.
      properties:
        col:
          type: string
        op:
          type: string
          enum: [eq, in, between, gte, lte]
        value:
          oneOf:
            - type: string
            - type: number
          nullable: true
        values:
          type: array
          nullable: true
          items:
            type: string
        min:
          type: number
          nullable: true
        max:
          type: number
          nullable: true
      required:
        - col
        - op
    TimeRange:
      type: object
      properties:
        kind:
          type: string
          enum: [ltm, year, ytd]
        year:
          type: integer
          nullable: true
      required:
        - kind
    QueryRequest:
      type: object
      properties:
        dataset:
          type: string
          enum: [revenue, kpi]
        metrics:
          type: array
          items:
            type: string
        group_by:
          type: array
          nullable: true
          items:
            type: string
            enum: [company, store, month, quarter, year, kpi_code]
        filters:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/QueryFilter"
        time_range:
          allOf:
            - $ref: "#/components/schemas/TimeRange"
          nullable: true
        limit:
          type: integer
          nullable: true
      required:
        - dataset
        - metrics
    QueryColumn:
      type: object
      properties:
        key:
          type: string
        label_ko:
          type: string
        format:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
      required:
        - key
        - label_ko
    QueryMeta:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        cache_hit:
          type: boolean
      required:
        - generated_at
        - cache_hit
    QueryResponse:
      type: object
      properties:
        columns:
          type: array
          items:
            $ref: "#/components/schemas/QueryColumn"
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        meta:
          $ref: "#/components/schemas/QueryMeta"
      required:
        - columns
        - rows
        - meta
    QueryError:
      type: object
      properties:
        error:
          type: string
      required:
        - error