# lib/api-zod/src/generated/types/financialStatementPreview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FinancialStatementPreviewStatementType } from "./financialStatementPreviewStatementType";
import type { ParsedFinancialLine } from "./parsedFinancialLine";
import type { StatementVerifyIssue } from "./statementVerifyIssue";

export interface FinancialStatementPreview {
  statement_type: FinancialStatementPreviewStatementType;
  line_count: number;
  lines: ParsedFinancialLine[];
  issues: StatementVerifyIssue[];
}