# lib/api-zod/src/generated/types/financialStatementPreviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FinancialStatementPreviewRequestStatementType } from "./financialStatementPreviewRequestStatementType";

export interface FinancialStatementPreviewRequest {
  markdown: string;
  statement_type: FinancialStatementPreviewRequestStatementType;
}