# lib/api-zod/src/generated/types/financialStatementSaveError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatementVerifyIssue } from "./statementVerifyIssue";

export interface FinancialStatementSaveError {
  error: string;
  issues: StatementVerifyIssue[];
}