# lib/api-zod/src/generated/types/financialStatementSaveRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FinancialStatementSaveRequestStatementType } from "./financialStatementSaveRequestStatementType";

export interface FinancialStatementSaveRequest {
  company_id: string;
  fiscal_year: number;
  statement_type: FinancialStatementSaveRequestStatementType;
  markdown: string;
  period_start?: string | null;
  period_end?: string | null;
  skip_verification?: boolean | null;
}