# lib/api-zod/src/generated/types/financialStatementSaveResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatementVerifyIssue } from "./statementVerifyIssue";

export interface FinancialStatementSaveResult {
  statement_id: string;
  line_count: number;
  issues: StatementVerifyIssue[];
}