# lib/api-zod/src/generated/types/parsedFinancialLine.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ParsedFinancialLine {
  account_name_ko: string;
  section_code?: string | null;
  depth: number;
  is_subtotal: boolean;
  amount?: string | null;
  prior_amount?: string | null;
}