# lib/api-zod/src/generated/types/statementVerifyIssue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatementVerifyIssuePeriod } from "./statementVerifyIssuePeriod";

export interface StatementVerifyIssue {
  period: StatementVerifyIssuePeriod;
  message: string;
}